requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0